jumiascraper/ml/ml_requirements.txt
pandas
matplotlib
seaborn
python-dotenv
scikit-learn
